+++ CHANGES.md
@@ -1,3 +1,8 @@
+### v0.9.0
+
+ - adds Break(key,value,...) for temporary runtime inspection. 
+ - add optional filename to Dump()
+
 ### v0.8.1
 
  - Explore replaces object with same root label.